[REFACTOR]: S3 관련 코드 인프라 패키지로 이동
