+++ src/main/java/project/domain/image/api/ImageController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import project.domain.image.dto.ImageUploadResponse;
 import project.domain.image.service.ImageService;
-import project.domain.image.service.S3Service;
+import project.infra.s3.application.S3Service;
 
 @RestController
 @RequiredArgsConstructor
